decouple Select stories
